# packages/tesslate-marketplace/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Tesslate Federated Marketplace `/v1`
  version: v1
  description: |
    Federated marketplace wire protocol for OpenSail / Tesslate Studio.

    Every response includes `X-Tesslate-Hub-Id` and
    `X-Tesslate-Hub-Api-Version: v1` headers. Endpoints whose capability
    isn't advertised in `/v1/manifest.capabilities[]` return a typed
    `501 unsupported_capability` JSON envelope.
  contact:
    name: Tesslate
    email: marketplace@tesslate.com
servers:
  - url: http://localhost:8800
    description: Local dev
  - url: https://marketplace.tesslate.com
    description: Tesslate Official
...
paths:
...
  /v1/yanks:
    get:
      tags: [yanks]
      summary: Yank-only subset of the changes feed
      parameters:
        - in: query
          name: since
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Yank events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangesFeed"
    post:
      tags: [yanks]
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YankCreate"
      responses:
        "201":
          description: Yank created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Yank"
  /v1/yanks/{id}:
    get:
      tags: [yanks]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Yank detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Yank"
  /v1/yanks/{id}/appeal:
    post:
      tags: [yanks]
      security: [{ bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YankAppealCreate"
      responses:
        "201":
          description: Appeal recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YankAppeal"
  /v1/items/{kind}/{slug}/reviews:
    get:
      tags: [reviews]
      parameters:
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewList"
    post:
      tags: [reviews]
      security: [{ bearer: [] }]
      parameters:
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Slug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /v1/items/{kind}/{slug}/reviews/aggregate:
    get:
      tags: [reviews]
      parameters:
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Aggregate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewAggregate"
...
  /v1/telemetry/install:
    post:
      tags: [telemetry]
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelemetryEvent"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetryAck"
  /v1/telemetry/usage:
    post:
      tags: [telemetry]
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelemetryEvent"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetryAck"
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    Kind:
      in: path
      name: kind
      required: true
      schema:
        type: string
        enum: [agent, skill, mcp_server, base, app, theme, workflow_template]
    Slug:
      in: path
      name: slug
      required: true
      schema: { type: string }
...
  schemas:
...
    ChangeEvent:
      type: object
      required: [op, kind, slug, etag]
      properties:
        op:
          type: string
          enum: [upsert, delete, deactivate, yank, version_remove, pricing_change]
        kind: { type: string }
        slug: { type: string }
        version: { type: string, nullable: true }
        etag: { type: string }
        payload: { type: object, additionalProperties: true, nullable: true }
        created_at: { type: string, format: date-time }
    ChangesFeed:
      type: object
      properties:
        events:
          type: array
          items: { $ref: "#/components/schemas/ChangeEvent" }
        next_etag: { type: string }
        has_more: { type: boolean }
    Review:
      type: object
      properties:
        id: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        title: { type: string, nullable: true }
        body: { type: string, nullable: true }
        reviewer_handle: { type: string }
        reviewer_avatar_url: { type: string, nullable: true }
        is_verified_install: { type: boolean }
        created_at: { type: string, format: date-time }
    ReviewList:
      type: object
      properties:
        reviews:
          type: array
          items: { $ref: "#/components/schemas/Review" }
        next_cursor: { type: string, nullable: true }
        has_more: { type: boolean }
    ReviewAggregate:
      type: object
      properties:
        count: { type: integer }
        mean: { type: number }
        distribution:
          type: object
          additionalProperties: { type: integer }
    ReviewCreate:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        title: { type: string, nullable: true }
        body: { type: string, nullable: true }
        reviewer_handle: { type: string, nullable: true }
...
    YankCreate:
      type: object
      required: [kind, slug, reason]
      properties:
        kind: { type: string }
        slug: { type: string }
        version: { type: string, nullable: true }
        severity:
          type: string
          enum: [low, medium, critical]
        reason: { type: string }
        requested_by: { type: string, nullable: true }
    Yank:
      type: object
      properties:
        id: { type: string }
        kind: { type: string }
        slug: { type: string }
        version: { type: string, nullable: true }
        severity: { type: string }
        reason: { type: string }
        requested_by: { type: string, nullable: true }
        state: { type: string }
        resolved_at: { type: string, format: date-time, nullable: true }
        resolution: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    YankAppealCreate:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
        submitted_by: { type: string, nullable: true }
    YankAppeal:
      type: object
      properties:
        id: { type: string }
        yank_id: { type: string }
        submitted_by: { type: string, nullable: true }
        reason: { type: string }
        state: { type: string }
        decision: { type: string, nullable: true }
        decision_reason: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TelemetryEvent:
      type: object
      required: [kind, slug]
      properties:
        kind: { type: string }
        slug: { type: string }
        version: { type: string, nullable: true }
        install_id: { type: string, nullable: true }
        payload: { type: object, additionalProperties: true, nullable: true }
    TelemetryAck:
      type: object
      properties:
        accepted: { type: boolean }
        received_at: { type: string, format: date-time }